+++ doc.go
@@ -1,0 +1,13 @@
+// Package app contains core structures for building flotilla applications:
+// App, Configuration, and Environment. Additionally, package app provides
+// management of key application functions through Environment:
+// determination of operation mode, creation of state, logging, and management
+// of extensions.
+// App
+// Configuration
+// Environment
+//  - Mode
+//  - State
+//  - Logging
+//  - Extensions
+package app